Compute the average of N runs (class inat.analyser.uppaal.ResultAverager), and show the resulting graph, with standard deviation as error bars (not very nice to see: have still to work on that part)
git-svn-id: http://hmisvn.ewi.utwente.nl/ExBio/projects/inat@55 d52acf90-180f-11e0-a790-6b35089b84c1

+++ src/inat/analyser/LevelResult.java
@@ -20,7 +20,7 @@ public interface LevelResult {
 	 * @param time the time index to do a look up for
 	 * @return the level of concentration
 	 */
-	public int getConcentration(String id, int time);
+	public double getConcentration(String id, double time);
 
 	/**
 	 * Determines the reactant ID's of substrates of which result are known.
@@ -34,7 +34,7 @@ public interface LevelResult {
 	 * 
 	 * @return the list of data point time indices
 	 */
-	public List<Integer> getTimeIndices();
+	public List<Double> getTimeIndices();
 	
 	public boolean isEmpty();
 }
